+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+- [FEATURE] Add evaluation logging to `DatadogFlags` module. See [#2646][]
 - [FEATURE] Automatic network instrumentation now tracks `URLSession` requests without requiring delegate registration. See [#2620][]
 - [FEATURE] Deprecate `URLSessionInstrumentation.enable(with:in:)` API in favor of `URLSessionInstrumentation.enableDurationBreakdown(with:in:)`. See [#2634][]
 - [IMPROVEMENT] Skip malformed Logs attributes individually instead of dropping the entire event, and log clear error messages. See [#2665][]
@@ -8,7 +9,7 @@
 # 3.6.1 / 02-02-2026
 
 - [FIX] Prevent crashes related to swapping the `__cxa_throw` function. See [#2661][]
-        
+
 # 3.6.0 / 28-01-2026
 
 - [FEATURE] Add `DatadogProfiling` module to profile app launches. See [#2654][]
@@ -1042,6 +1043,7 @@ Release `2.0` introduces breaking changes. Follow the [Migration Guide](MIGRATIO
 [#2634]: https://github.com/DataDog/dd-sdk-ios/pull/2634
 [#2639]: https://github.com/DataDog/dd-sdk-ios/pull/2639
 [#2640]: https://github.com/DataDog/dd-sdk-ios/pull/2640
+[#2646]: https://github.com/DataDog/dd-sdk-ios/pull/2646
 [#2647]: https://github.com/DataDog/dd-sdk-ios/pull/2647
 [#2654]: https://github.com/DataDog/dd-sdk-ios/pull/2654
 [#2661]: https://github.com/DataDog/dd-sdk-ios/pull/2661